Struct mazelib::implm::coordinate::inline::BoxSpaceInlineCellMazeCoordinator

pub struct BoxSpaceInlineCellMazeCoordinator<Buffer: MazeBuffer<InlineCellValue<DIMENSION>>, const DIMENSION: usize> { /* private fields */ }

TODO

As this coordinator implements a one-to-one mapping between points and cells, there is no separate CellLocation struct. CoordinateTuplets are converted directly into CellIDs.

Implementations

impl<Buffer: MazeBuffer<InlineCellValue<DIMENSION>>, const DIMENSION: usize> BoxSpaceInlineCellMazeCoordinator<Buffer, DIMENSION>

pub fn buffer(&self) -> &Buffer

pub fn set( &mut self, cell: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType, value: <Self as MazeCoordinator>::CellVal )

Set the value of any cell, including ones not mapped by the coordinate space

Since with a InlineCellValue it is impossible to get the maze into an inconsistent state, this function is not considered unsafe.

In most cases you should use the methods on MazeCoordinator instead of this.