Trait mazelib::implm::export::minecraft::SchematicMazeExporter

pub trait SchematicMazeExporter<M: MazeCoordinator, O: Write>: MazeExporter<M, O> { }

Available on crate feature minecraft only.

Export a 2D maze into a Minecraft schematic.

Specifically, it exports the maze to a Sponge Schematic (v3), and returns it as a gzipped serialised NBT byte array. Note that schematics cannot be used directly with the vanilla game, and must be imported with mods or 3rd-party tools (such as WorldEdit).

The blocks used cannot be changed, however most programs that let you import schematics also let you replace blocks of one type with another. The vanilla game also supports this on small areas with /fill <from> <to> replace <block>.