Struct mazelib::implm::export::text::BoxSpaceInlineCellTextMazeExporter

pub struct BoxSpaceInlineCellTextMazeExporter { /* private fields */ }

An exporter that renders BoxSpaceInlineCellMazeCoordinators to text.

Implementations

impl BoxSpaceInlineCellTextMazeExporter

pub fn builder() -> BoxSpaceInlineCellTextMazeExporterBuilder

Construct a new builder for a BoxSpaceInlineCellTextMazeExporter.

Optional, see DefaultMazeExporter.

pub fn chars_per_cell_horizontally(&self) -> NonZeroUsize

Return the number of the characters that is emitted horizontally per cell, excluding walls.

pub fn chars_per_cell_vertically(&self) -> NonZeroUsize

Return the number of the characters that is emitted vertically per cell, excluding walls.