Struct mazelib::implm::maze::block::BoxSpaceBlockCellMaze

pub struct BoxSpaceBlockCellMaze<Buffer: MazeBuffer<BlockCellValue>, const DIMENSION: usize> { /* private fields */ }

TODO write a description

Examples

With scaling and padding:

use mazelib::implm::point::boxy::TwoDimensionalBoxCoordinateSpace;
use mazelib::implm::cell::block::{BlockCellValue, BoxSpaceBlockCellManagerBuilder};
use mazelib::implm::buffer::VecBuffer;

let coord_space = TwoDimensionalBoxCoordinateSpace::new([11, 11]); // Standard 2D coordinate space
let scale_factor = [2, 2]; // A simple scale factor of 2 for a clean look
let padding = [[1, 1], [1, 1]]; // 1 cell on each side for a border
type CellType = BlockCellValue; // Pixelated maze style

type MazeBuffer = VecBuffer<CellType>;

let maze = BoxSpaceBlockCellManagerBuilder::<MazeBuffer, 2>::new(coord_space)
                                            .scale_factor(scale_factor)
                                            .padding(padding)
                                            .build();

Implementations

impl<Buffer: MazeBuffer<BlockCellValue>, const DIMENSION: usize> BoxSpaceBlockCellManager<Buffer, DIMENSION>

pub fn buffer(&self) -> &Buffer

pub fn get_full_dimensions(&self) -> [usize; DIMENSION]

The dimensions of the coordinate space, scaled by the resolution, plus padding.

pub fn scale_factor(&self) -> [usize; DIMENSION]

The number of cells between points, minus one.

A scale factor of 1 would have each point directly adjacent to one another. A scale factor of 2 will have 1 cell between each point.

Does not affect the distance of points from the outer edge of the maze (see padding()).

pub fn padding(&self) -> [[usize; 2]; DIMENSION]

The number of cells between the edge of the maze and the outermost cell that is mapped to a point. Useful for borders when the scale factor is greater than one.

pub fn scale_dimensions(
dimensions: [usize; DIMENSION],
resolution: [usize; DIMENSION]
) -> [usize; DIMENSION]

The dimensions of the coordinate space, scaled by the resolution

pub fn map_pt_to_cell_loc(
&self,
pt: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
) -> BlockCell<DIMENSION>

Map a point to a cell location.

pub fn get_mut(
&mut self,
pt: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
) -> &mut <Self as CellManager>::CellVal

Get the value of the point pt for mutation.

See also: Self::get().

pub fn set_type(
&mut self,
pt: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType,
cell_type: BlockCellValueType
)

Sugar for

self.get_mut(pt).cell_type = cell_type;

Modifies the cell’s type but keeps its mark status intact.

Please see Self::get_mut() for details.

pub fn get_cell_value(&self, loc: BlockCell<DIMENSION>) -> BlockCellValue

Get the value of any cell.

In most cases you should use the methods on CellManager::get(). The only reason you should use this method is to access a cell that is not mapped by the coordinated space.

pub fn get_cell_value_mut(
&mut self,
loc: BlockCell<DIMENSION>
) -> &mut BlockCellValue

Get the value of any cell for mutation.

In most cases you should use the methods on Self::get_mut(). The only reason you should use this method is to access a cell that is not mapped by the coordinated space.

See also: Self::get_cell_value().

pub fn set_cell_value(
&mut self,
loc: BlockCell<DIMENSION>,
value: BlockCellValue
)

Set the value of any cell, including ones not mapped by the coordinate space

Since with a BlockCellValue it is impossible to get the maze into an inconsistent state, this function is not considered unsafe.

In most cases you should use the methods on CellManager. The only reason you should use this method is to access a cell that is not mapped by the coordinated space.

pub fn set_cell_value_type(
&mut self,
loc: BlockCell<DIMENSION>,
cell_type: BlockCellValueType
)

Sugar for

self.get_cell_value_mut(loc).cell_type = cell_type;

Please see Self::get_cell_value_mut() for details.