Module mazelib::implm::point::boxy

Box coordinate spaces. This includes square mazes, rectangular mazes, cuboid mazes, and other hyperrectangular constructions.

Commonly used coordinate spaces:

TwoDimensionalBoxCoordinateSpace

ThreeDimensionalBoxCoordinateSpace

Dimensions

x always represents the “lowest” dimension. A one-dimensional point’s only has an x dimension, a two-dimensional point has an x dimension and a y dimension.

The x dimension represents columns, the y dimension represents rows, the z dimension represents layers, and so on.

Order of Dimensions

In this crate, the most minor dimension always comes first. For example, you would index a CoordinateTuplet could be represented as [x, y, z], not [z, y, x]. Likewise, a coordinate space for a maze with three layers, four rows, and five columns could be constructed as:

BoxCoordinateSpace::new([5, 4, 3]);

You could also use ThreeDimensionalBoxCoordinateSpace in this example instead of BoxCoordinateSpace — they would be equivalent.

Structs

BoxCoordinateSpace

An n-dimensional coordinate space shaped like a box.

CoordinateTuplet

A tuplet of unsigned integers that uniquely represent a point in n-dimensional space.

Type Definitions

CoordinatePair

An (x, y) pair for referencing points in a TwoDimensionalBoxCoordinateSpace.

CoordinateTriplet

An (x, y, z) triplet for referencing points in a ThreeDimensionalBoxCoordinateSpace.

ThreeDimensionalBoxCoordinateSpace

A three dimensional rectangular cuboid coordinate space.

TwoDimensionalBoxCoordinateSpace

A flat rectangular coordinate space. The most common type of coordinate space for mazes.