Struct mazelib::implm::point::boxy::BoxCoordinateSpace

pub struct BoxCoordinateSpace<const DIMENSION: usize> { /* private fields */ }

An n-dimensional coordinate space shaped like a box with box-like points.

It is a generalisation of the common two-dimensional and three-dimensional Cartesian plane/space to n dimensions.

All (size) dimensions of the space must be at least 1. The (mathematical) dimension (DIMENSION) must be also be at least 1.

The origin of a boxy coordinate space is at the point where every coordinate is zero.

Coordinate Axes

As a BoxCoordinateSpace has n dimensions, there are n coordinate axes. These are labelled from 0 to n - 1. The most minor axis is 0, and the most major axis is n - 1. These correspond directly with the array or tuple indices of a coordinate tuplet.

Adjacency

Direct Adjacency

Two points are considered directly adjacent if:

On one single coordinate axis, their respective coordinates differ by exactly 1, and,

On every other axis, their respective coordinates are the same.

For example, (1, 1, 1) and (1, 2, 1) are considered directly adjacent. (1, 1, 1) and (1, 2, 2) are not directly adjacent. (1, 1, 1) and (1, 1, 1) are also not considered directly adjacent.

Edge Adjacency

A point is considered to be adjacent to the edge of the coordinate space if there is some axis where the point’s respective coordinate is 0 or the length of axis - 1.

Implementations

impl<const DIMENSION: usize> BoxCoordinateSpace<DIMENSION>

pub fn new(dimensions: [NonZeroUsize; DIMENSION]) -> Self

Construct a new BoxCoordinateSpace from the given (size) dimensions.

Parameters

dimensions — the width, height, depth, etc. of the coordinate space. As the (mathematical) dimension must be non-zero, this array cannot be empty. Additionally, the product of all dimensions must fit within a usize. (Mazes this large won’t fit in memory anyway).

Examples

// If you're using hard-coded constants like in this example,
// you may prefer new_checked() for the ergonomics
let coord_space = BoxCoordinateSpace::new([NonZeroUsize::new_unchecked(5), NonZeroUsize::new_unchecked(7)]);

See Also

new_checked()

pub fn new_checked(dimensions: [usize; DIMENSION]) -> Self

Construct a new BoxCoordinateSpace from the given (size) dimensions.

Parameters

dimensions — the width, height, depth, etc. of the coordinate space. As the (mathematical) dimension must be non-zero, this array cannot be empty. All dimensions must be non-zero as well. Additionally, the product of all dimensions must fit within a usize. (Mazes this large won’t fit in memory anyway).

Examples

let coord_space = BoxCoordinateSpace::new_checked([5, 7, 3]);

pub fn dimensions(&self) -> [NonZeroUsize; DIMENSION]

Return the (size) dimensions of this coordinate space.

pub fn is_adjacent_to_edge(&self, pt: <Self as CoordinateSpace>::PtType) -> bool

Return whether the point is adjacent to the edge of this coordinate space.