Type Alias mazelib::implm::point::boxy::CoordinateTriplet

pub type CoordinateTriplet = CoordinateTuplet<3>;

An (x, y, z) triplet for referencing points in a ThreeDimensionalBoxCoordinateSpace.

The x-coordinate represents the column, the y-coordinate represents the row, and the z-coordinate the layer.

Aliased Type

struct CoordinateTriplet(pub [usize; 3]);

Fields
0: [usize; 3]
Implementations

impl<const DIMENSION: usize> CoordinateTuplet<DIMENSION>

pub fn offset(&self, axis: usize, offset: isize) -> Self

Get the point at a given offset from this point along the given axis.

Examples

use mazelib::implm::point::boxy::CoordinateTuplet;

let pt = CoordinateTuplet([2, 2, 2]);

assert_eq!(CoordinateTuplet([2, 5, 2]), pt.offset(1, 3));
assert_eq!(CoordinateTuplet([1, 2, 2]), pt.offset(0, -1));

pub fn at(&self, axis: usize, position: usize) -> Self

Replace the position along the given axis with a new value.

Useful when iterating along an axis.

Suggestions for better names are welcome.

Examples

use mazelib::implm::point::boxy::CoordinateTuplet;

let pt = CoordinateTuplet([1, 1, 1]);

assert_eq!(CoordinateTuplet([1, 1, 0]), pt.at(2, 0));

pub fn any_coordinate_is(&self, value: usize) -> bool

Return true if any coordinate of this tuplet is the same as value.

Examples

use mazelib::implm::point::boxy::CoordinateTuplet;

assert!(CoordinateTuplet([7, 8, 9]).any_coordinate_is(8));
assert!(CoordinateTuplet([0, 0, 1]).any_coordinate_is(0));

assert_eq!(false, CoordinateTuplet([7, 8, 9]).any_coordinate_is(1));
assert_eq!(false, CoordinateTuplet([0, 0, 1]).any_coordinate_is(2));

pub fn axis_of_adjacency_with( &self, other_pt: CoordinateTuplet<DIMENSION> ) -> Option<usize>

Return the axis upon which this point is directly adjacent with another point.

Examples

use mazelib::implm::point::boxy::CoordinateTuplet;

assert_eq!(Some(1), CoordinateTuplet([5, 5, 5]).axis_of_adjacency_with(CoordinateTuplet([5, 4, 5])));
assert_eq!(Some(0), CoordinateTuplet([14, 2, 8]).axis_of_adjacency_with(CoordinateTuplet([13, 2, 8])));
assert_eq!(None, CoordinateTuplet([2, 2, 2]).axis_of_adjacency_with(CoordinateTuplet([2, 3, 3])));
assert_eq!(None, CoordinateTuplet([1, 1, 1]).axis_of_adjacency_with(CoordinateTuplet([1, 1, 1])));