Type Alias mazelib::implm::point::boxy::ThreeDimensionalBoxCoordinateSpace

pub type ThreeDimensionalBoxCoordinateSpace = BoxCoordinateSpace<3>;

A three dimensional rectangular cuboid coordinate space.

Like TwoDimensionalBoxCoordinateSpace, but 3D.

While this is optimised for cuboid mazes, you can use whatever maze outline you want. (You can create custom outlines by mutating the maze before passing it into a generator).

Aliased Type

struct ThreeDimensionalBoxCoordinateSpace { /* private fields */ }

Implementations

impl<const DIMENSION: usize> BoxCoordinateSpace<DIMENSION>

pub fn new(dimensions: [NonZeroUsize; DIMENSION]) -> Self

Construct a new BoxCoordinateSpace from the given (size) dimensions.

Parameters

dimensions — the width, height, depth, etc. of the coordinate space. As the (mathematical) dimension must be non-zero, this array cannot be empty. Additionally, the product of all dimensions must fit within a usize. (Mazes this large won’t fit in memory anyway).

Examples

// If you're using hard-coded constants like in this example,
// you may prefer new_checked() for the ergonomics
let coord_space = BoxCoordinateSpace::new([NonZeroUsize::new_unchecked(5), NonZeroUsize::new_unchecked(7)]);

See Also

new_checked()

pub fn new_checked(dimensions: [usize; DIMENSION]) -> Self

Construct a new BoxCoordinateSpace from the given (size) dimensions.

Parameters

dimensions — the width, height, depth, etc. of the coordinate space. As the (mathematical) dimension must be non-zero, this array cannot be empty. All dimensions must be non-zero as well. Additionally, the product of all dimensions must fit within a usize. (Mazes this large won’t fit in memory anyway).

Examples

let coord_space = BoxCoordinateSpace::new_checked([5, 7, 3]);

pub fn dimensions(&self) -> [NonZeroUsize; DIMENSION]

Return the (size) dimensions of this coordinate space.

pub fn is_adjacent_to_edge(&self, pt: <Self as CoordinateSpace>::PtType) -> bool

Return whether the point is adjacent to the edge of this coordinate space.