Module mazelib::interface::buffer

Working maze storage.

Mazes are typically stored in memory, but do not have to be. It is up to the implementation of the MazeBuffer.

For exporting mazes to persistent forms (like images), see super::render.

Recommended Reading

MazeBuffer – the buffer trait.

Traits

MazeBuffer

Maze cell container.