Module mazelib::interface::cell

Reading and manipulating mazes at a high level.

Mazes are fundamentally a collection of cells. These collections are stored in MazeBuffers.

While a collection of cells constitutes a maze, they need to be interpreted before any useful analysis can be done. This is the purpose of the CellManager - to hold the necessary information for interpreting cells.

Recommended Reading

crate::interface’s section on cells — a detailed introduction to cells and what they are.

ConnectionType — the ways that points can be connected.

CellManager — the unifying glue and centre of the maze model.

Structs

CellID

A unique identifier for a cell within a given maze.

Enums

ConnectionType

Type of connection (graph theory: edge) (e.g. wall, passage) between two points.

Traits

CellLocation

Location of a cell.

CellManager

Translates between points, cell locations, and cell identifiers. Executes queries on mazes and modifications to mazes.

CellValue

Value of a cell.