Trait mazelib::interface::cell::CellManager

pub trait CellManager: Debug {
    type CoordSpace: CoordinateSpace;
    type CellVal: CellValue;

    fn coord_space(&self) -> &Self::CoordSpace;
    fn get(
        &self, 
        pt: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
    ) -> Self::CellVal;
    fn get_connection(
        &self, 
        from: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType, 
        to: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
    ) -> ConnectionType;
    fn make_passage(
        &mut self, 
        pt: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
    );
    fn make_passage_between(
        &mut self, 
        from: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType, 
        to: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
    );
    fn make_wall(
        &mut self, 
        pt: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
    );
    fn make_wall_between(
        &mut self, 
        from: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType, 
        to: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
    );
    fn make_boundary(
        &mut self, 
        pt: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
    );
    fn make_boundary_between(
        &mut self, 
        from: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType, 
        to: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
    );

    fn is_passage_between(
        &self, 
        from: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType, 
        to: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
    ) -> bool { ... }
    fn is_wall_between(
        &self, 
        from: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType, 
        to: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
    ) -> bool { ... }
    fn is_boundary_between(
        &self, 
        from: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType, 
        to: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
    ) -> bool { ... }
    fn is_unvisited_between(
        &self, 
        from: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType, 
        to: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
    ) -> bool { ... }

}

Translates between points, cell locations, and cell identifiers. Executes queries on mazes and modifications to mazes.

Cells have no inherent meaning. They must be interpreted in context. This is exactly what a Cell Manager does.

Due to requiring intimate knowledge of how the coordinate space works and how the cell space works, each Cell Manager is specific to a single CoordinateSpace-CellValue pair. There is generally only one such manager for each pair.

Required Associated Types

type CoordSpace: CoordinateSpace

The type of coordinate space this cell manager supports.

type CellVal: CellValue

The value of the type of cell this cell manager uses.

Required Methods

fn coord_space(&self) -> &Self::CoordSpace

Return the coordinate space for this maze.

fn get(
&self,
pt: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
) -> Self::CellVal

Return the value of the point pt.

fn get_connection(
&self,
from: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType,
to: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
) -> ConnectionType

Return the type of connection (graph theory: edge) between two points.

If you attempted to walk from from to to this is what you would encounter.

The order of the arguments is important and has semantic meaning. Swapping the arguments may produce different results.

This method follows the priority order outlined in ConnectionType.

fn make_passage(
&mut self,
pt: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
)

Make a passage at pt.

fn make_passage_between(
&mut self,
from: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType,
to: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
)

Make a passage from from to to. The points must be adjacent. If they are not, this function will panic with debug assertions enabled. Passing in non-adjacent points without debug assertions enabled is undefined.

The order of the arguments is important and has semantic meaning. Swapping the arguments may produce different results.

fn make_wall(
&mut self,
pt: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
)

Make a wall at pt.

fn make_wall_between(
&mut self,
from: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType,
to: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
)

Make a wall from from to to. The points must be adjacent. If they are not, this function will panic with debug assertions enabled. Passing in non-adjacent points without debug assertions enabled is undefined.

The order of the arguments is important and has semantic meaning. Swapping the arguments may produce different results.

fn make_boundary(
&mut self,
pt: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
)

Make a boundary at pt.

fn make_boundary_between(
&mut self,
from: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType,
to: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
)

Make a boundary from from to to. The points must be adjacent. If they are not, this function will panic with debug assertions enabled. Passing in non-adjacent points without debug assertions enabled is undefined.

The order of the arguments is important and has semantic meaning. Swapping the arguments may produce different results.

Provided Methods

fn is_passage_between(
&self,
from: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType,
to: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
) -> bool

Return true when get_connection(from, to) returns ConnectionType::PASSAGE.

fn is_wall_between(
&self,
from: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType,
to: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
) -> bool

Return true when get_connection(from, to) returns ConnectionType::WALL.

fn is_boundary_between(
&self,
from: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType,
to: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
) -> bool

Return true when get_connection(from, to) returns ConnectionType::BOUNDARY.

fn is_unvisited_between(
&self,
from: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType,
to: <<Self as CellManager>::CoordSpace as CoordinateSpace>::PtType
) -> bool

Return true when get_connection(from, to) returns ConnectionType::UNVISITED.