Trait mazelib::interface::coordinate::MazeCoordinator

pub trait MazeCoordinator: Debug {
    type CoordSpace: CoordinateSpace;
    type CellLoc: CellLocation;
    type CellVal: CellValue;

    // Required methods
    fn coord_space(&self) -> &Self::CoordSpace;
    fn get(
        &self,
        pt: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType
    ) -> Self::CellVal;
    fn get_connection(
        &self,
        from: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType,
        to: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType
    ) -> ConnectionType;
    fn make_passage(
        &mut self,
        pt: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType
    );
    fn make_passage_between(
        &mut self,
        from: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType,
        to: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType
    );
    fn make_wall(
        &mut self,
        pt: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType
    );
    fn make_wall_between(
        &mut self,
        from: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType,
        to: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType
    );
    fn make_boundary(
        &mut self,
        pt: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType
    );
    fn make_boundary_between(
        &mut self,
        from: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType,
        to: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType
    );

    // Provided methods
    fn is_passage_between(
        &self,
        from: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType,
        to: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType
    ) -> bool { ... }
    fn is_wall_between(
        &self,
        from: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType,
        to: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType
    ) -> bool { ... }
    fn is_boundary_between(
        &self,
        from: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType,
        to: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType
    ) -> bool { ... }
    fn is_unvisited_between(
        &self,
        from: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType,
        to: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType
    ) -> bool { ... }

}

Handles high-level queries on mazes.

The coordinator defines the mapping between a maze’s coordinate space and its cell space. This means it understands how to translate high-level queries (like “What is the connection type between these points?”) into checks on specific cells in the maze’s buffer. It also understands which cells need to be modified and how to perform basic changes to a maze, like carving a new passageway.

The coordinator owns all of a maze’s components. This means that the MazeCoordinator effectively is the maze itself.

Each coordinator has different characteristics that dramatically affect the maze’s structure so you should consult your chosen implementation carefully.

Required Associated Types

type CoordSpace: CoordinateSpace

The type of coordinate space this coordinator supports.

type CellLoc: CellLocation

The location type for the cell type this coordinator uses.

type CellVal: CellValue

The value type for the cell type this coordinator uses.

Required Methods

fn coord_space(&self) -> &Self::CoordSpace

Return the coordinate space for this maze.

fn get( &self, pt: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType ) -> Self::CellVal

Return the value of the point pt.

fn get_connection( &self, from: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType, to: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType ) -> ConnectionType

Return the type of connection (graph theory: edge) between two points.

If you attempted to walk from from to to this is what you would encounter.

The order of the arguments is important and has semantic meaning. Swapping the arguments may produce different results.

This method follows the priority order outlined in ConnectionType.

fn make_passage( &mut self, pt: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType )

Make a passage at pt.

fn make_passage_between( &mut self, from: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType, to: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType )

Make a passage from from to to. The points must be adjacent. If they are not, this function will panic with debug assertions enabled. Passing in non-adjacent points without debug assertions enabled is undefined.

The order of the arguments is important and has semantic meaning. Swapping the arguments may produce different results.

fn make_wall( &mut self, pt: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType )

Make a wall at pt.

fn make_wall_between( &mut self, from: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType, to: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType )

Make a wall from from to to. The points must be adjacent. If they are not, this function will panic with debug assertions enabled. Passing in non-adjacent points without debug assertions enabled is undefined.

The order of the arguments is important and has semantic meaning. Swapping the arguments may produce different results.

fn make_boundary( &mut self, pt: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType )

Make a boundary at pt.

fn make_boundary_between( &mut self, from: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType, to: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType )

Make a boundary from from to to. The points must be adjacent. If they are not, this function will panic with debug assertions enabled. Passing in non-adjacent points without debug assertions enabled is undefined.

The order of the arguments is important and has semantic meaning. Swapping the arguments may produce different results.

Provided Methods

fn is_passage_between( &self, from: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType, to: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType ) -> bool

Return true when get_connection(from, to) returns ConnectionType::PASSAGE.

fn is_wall_between( &self, from: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType, to: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType ) -> bool

Return true when get_connection(from, to) returns ConnectionType::WALL.

fn is_boundary_between( &self, from: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType, to: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType ) -> bool

Return true when get_connection(from, to) returns ConnectionType::BOUNDARY.

fn is_unvisited_between( &self, from: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType, to: <<Self as MazeCoordinator>::CoordSpace as CoordinateSpace>::PtType ) -> bool

Return true when get_connection(from, to) returns ConnectionType::UNVISITED.