Trait mazelib::interface::export::MazeExporter

pub trait MazeExporter<M: MazeCoordinator, O: Write> {
    // Required method
    fn export(&self, maze: &M, output: &mut O) -> Result<()>;
}

Export a maze into another, usually persistent, format.

Maze coordinators are only useful within the confines of this library. If you want to visualise your handiwork you’ll want to export it to something you can see. If you want to use the maze in another program, you’ll want to write your own exporter that exports the maze to that program’s datastructures.

Exporters are wild and varied in their produce. What an exporter produces is completely up to them so make sure to consult your chosen implementation’s documentation.

Stability

For all implementations, unless explicitly documented otherwise, the exact details of the output should not be depended upon. For example, the implementation that exports to text may change the characters it outputs or the spacing. Generally you should not be perform any operations on an exported maze that can be done before export. If for some reason you do need the output to never change, just copy-paste the exporter into your own project (license-permitting). Or just never update the library.

Examples

Assuming you already have a maze:

let maze: BoxSpaceInlineCellMazeCoordinator<VecBuffer<InlineCellValue<2>>, 2> /* = ... */;

Write a maze to stdout:

use std::io::BufWriter;
use mazelib::implm::export::text::BoxSpaceInlineCellTextMazeExporter;
use mazelib::interface::export::MazeExporter;

// Buffer writes for performance
let mut output = BufWriter::new(std::io::stdout());

BoxSpaceInlineCellTextMazeExporter::builder().build().export(&maze, &mut output)

Write a maze to a file:

use std::fs::File;
use std::io::BufWriter;
use mazelib::implm::export::text::BoxSpaceInlineCellTextMazeExporter;
use mazelib::interface::export::MazeExporter;

// Buffer writes for performance
let mut output = BufWriter::new(File::create("maze.txt")?);

BoxSpaceInlineCellTextMazeExporter::builder().build().export(&maze, &mut output)?;

Required Methods

fn export(&self, maze: &M, output: &mut O) -> Result<()>

Export the maze maze to writer output.

The format the maze is exported to (and thus the type of data that is written to output) is defined by and dependent on the implementation.

Returns an IO error if one was encountered while writing to output.

Usage Notes

You may wish to wrap output in a BufWriter if it not already buffered.