Trait mazelib::interface::point::CoordinateSpace

pub trait CoordinateSpace: Sized + Clone + Copy + Send + Sync + Debug {
    type PtType: Point;
    type Iter: Iterator<Item = Self::PtType>;

    // Required methods
    fn logical_size(&self) -> NonZeroUsize;
    fn neighbours_of_pt(&self, pt: Self::PtType) -> Vec<Self::PtType>;
    fn are_adjacent(&self, pt1: Self::PtType, pt2: Self::PtType) -> bool;
    fn iter(&self) -> Self::Iter;
    fn iter_from(&self, pt: Self::PtType) -> Self::Iter;
    fn choose(&self, rng: &mut (impl Rng + ?Sized)) -> Self::PtType;
}

A mathematical graph that represents all possible meaningful positions in a maze.

It defines where potential junctions are and how they connect to other potential junctions. It does not encode any information about passages or walls, only the locations where they could go.

For a more in-depth explanation, see the explainer in interface.

A coordinate space must always have at least one point. It can never be empty.

Examples

Constructing a maze with a specific coordinate space:

#![feature(generic_arg_infer)]  // Optional, lets the compiler infer whether it's 2D, 3D, etc.

let dimensions = [7, 7];

let space = BoxCoordinateSpace::new_checked(dimensions);

// Will need to pick your own buffer
let maze = BoxSpaceBlockCellMazeCoordinator::<Buffer, _>::builder(space).build();

See Also

Point – the individual parts of a coordinate space

Required Associated Types

type PtType: Point

The Point type that goes with this coordinate space.

type Iter: Iterator<Item = Self::PtType>

The type of iterator returned by iter() and iter_from().

Required Methods

fn logical_size(&self) -> NonZeroUsize

Return the number of points in this coordinate space.

This value is fixed.

As the coordinate space must always have at least one point, this method returns NonZeroUsize.

Examples

// A 3x3 space has 9 points
let space = BoxCoordinateSpace::new_checked([3, 3]);

assert_eq!(9, usize::from(space.logical_size()));

fn neighbours_of_pt(&self, pt: Self::PtType) -> Vec<Self::PtType>

Return a vector containing every point in the coordinate space that is adjacent to pt (as determined by are_adjacent()).

Examples

let space = BoxCoordinateSpace::new_checked([3, 3]);

/*
 * Point = X
 * Neighbours = *
 *
 * ┌───┬───┬───┐
 * │   │   │   │
 * ├───┼───┼───┤
 * │   │ * │   │
 * ├───┼───┼───┤
 * │ * │ X │ * │
 * └───┴───┴───┘
 */

let pt = (1, 0).into();

let expected_neighbours: Vec<CoordinatePair> = vec![
    (0, 0).into(),
    (2, 0).into(),
    (1, 1).into()
];

assert_eq!(expected_neighbours, space.neighbours_of_pt(pt));

fn are_adjacent(&self, pt1: Self::PtType, pt2: Self::PtType) -> bool

Return whether two points are adjacent in this coordinate space.

A point is not adjacent to itself.

Examples

let space = BoxCoordinateSpace::new_checked([3, 3]);

// Adjacent
assert!(space.are_adjacent((1, 0).into(), (0, 0).into()));
assert!(space.are_adjacent((1, 0).into(), (1, 1).into()));
assert!(space.are_adjacent((1, 0).into(), (2, 0).into()));

// Not adjacent
assert!(!space.are_adjacent((1, 0).into(), (0, 1).into()));
assert!(!space.are_adjacent((1, 0).into(), (2, 2).into()));
assert!(!space.are_adjacent((1, 0).into(), (1, 2).into()));  // No wrapping

fn iter(&self) -> Self::Iter

Return an iterator that yields every point in this coordinate space.

Every point yielded is adjacent to a point previously yielded. The one exception to this is where there is no possible path from the current point to any previously yielded point. For example, the first point, or if the space contains two or more isolated components, the first point of each component.

If the coordinate space has any notion of an origin, it is the first point yielded.

Examples

let space = BoxCoordinateSpace::new_checked([2, 2]);

let mut iter = space.iter();

assert_eq!(Some((0, 0).into()), iter.next());
assert_eq!(Some((1, 0).into()), iter.next());
assert_eq!(Some((0, 1).into()), iter.next());
assert_eq!(Some((1, 1).into()), iter.next());

assert_eq!(None, iter.next());

fn iter_from(&self, pt: Self::PtType) -> Self::Iter

Return an iterator that behaves exactly like iter(), but skips every point up to and including pt.

Points that are skipped are still considered to have been yielded for the purposes of the path constraint in iter().

Examples

Contrast this example with the example for iter().

let space = BoxCoordinateSpace::new_checked([2, 2]);

let mut iter = space.iter_from((1, 0).into());

assert_eq!(Some((0, 1).into()), iter.next());
assert_eq!(Some((1, 1).into()), iter.next());

assert_eq!(None, iter.next());

fn choose(&self, rng: &mut (impl Rng + ?Sized)) -> Self::PtType

Return a random point in this coordinate space.

Examples

let mut rng = thread_rng();

let space = BoxCoordinateSpace::new_checked([3, 3]);

let pt = space.choose(&mut rng);