Trait mazelib::interface::point::Point

pub trait Point: Sized + Clone + Copy + PartialEq + Eq + Hash + Debug { }

A logical location in a maze.

In less abstract terms, points are basically the potential junctions in a maze.

Unlike cells, have semantic meaning, and it is the connection of points (by the CoordinateSpace) that determine the logical structure of the maze.

A detailed introduction can be found in the section on points in the core maze components explanation.

See Also

CoordinateSpace – the parent container of points.

CellLocation – physical (rather than logical) locations in a maze.

Implementation

With regards to implementation, points are or similar to coordinate tuplets (e.g. (x, y)). Points do not have to be coordinates, but must effectively act like them. More specifically, they must have the following properties:

Immutable

Fixed Size

Cloneable

Equality must be reflexive, symmetric, transitive, and anti-symmetric

Hashable