Trait mazelib::interface::render::MazeRenderer

pub trait MazeRenderer<M: Maze> {
    fn render<Output: Write + Seek>(
        &self, 
        maze: &M, 
        output: &mut Output
    ) -> Result<()>;
}

Render (export) a maze into another format.

Render a maze into another, usually non-volatile format. Each implementation of this trait will render mazes into different formats.

Stability

For all implementations, unless explicitly documented otherwise, the exact formatting of the render should not be depended upon. For example, the implementation that renders to text may change the characters it outputs or the spacing.

Generally you should not be perform any operations on a rendered maze that can be done before rendering.

Examples

Assuming you already have a maze:

let maze: BoxSpaceInlineCellManager<VecBuffer<InlineCellValue<2>>, 2> /* = ... */;

Write a maze to stdout:

use std::io::BufWriter;
use mazelib::implm::render::text::BoxSpaceTextMazeRenderer;
use mazelib::interface::render::MazeRendererNonSeeking;

// Buffer writes for performance
let mut output = BufWriter::new(std::io::stdout());

BoxSpaceTextMazeRenderer::new().render(&maze, &mut output)

Write a maze to a file:

use std::fs::File;
use std::io::BufWriter;
use mazelib::implm::render::text::BoxSpaceTextMazeRenderer;
use mazelib::interface::render::MazeRendererNonSeeking;

// Buffer writes for performance
let mut output = BufWriter::new(File::create("maze.txt")?);

BoxSpaceTextMazeRenderer::new().render(&maze, &mut output)?;

Implementing

If an implementation does not require a writer to support seeking, it should implement MazeRendererNonSeeking instead. Implementing that trait will also automatically implement this one.

Required Methods

fn render<Output: Write + Seek>(
&self,
maze: &M,
output: &mut Output
) -> Result<()>

Render the maze maze to writer output.

The format the maze is rendered to (and thus the type of data that is written to output) is defined by and dependent on the implementation.

Returns an IO error if one was encountered while writing to output.

Usage Notes

It is recommended to wrap output in a BufWriter if it not already buffered.