Trait mazelib::interface::render::MazeRendererNonSeeking

pub trait MazeRendererNonSeeking<M: Maze>: MazeRenderer<M> {
    fn render<Output: Write>(&self, maze: &M, output: &mut Output) -> Result<()>;
}

Like MazeRenderer, but supports writers that do not support seeking.

See MazeRenderer for details on this trait.

If you have a writer that requires this trait, make sure you import this trait and not MazeRenderer.

Required Methods

fn render<Output: Write>(&self, maze: &M, output: &mut Output) -> Result<()>