Struct mazelib::path::Path

pub struct Path<T: Sized + Clone + Copy + PartialEq + Eq + Hash + Send + Sync + Debug> { /* private fields */ }

A series of movements from location to location.

It is a directed path, so locations can only be traversed in order.

Examples

// Create a path starting at the "origin"
let mut path = Path::starting_at(0);
 
// Wander about
path.push(2);
path.push(3);
path.push(4);
path.push(2);
path.push(4);
path.push(3);
 
// Remove loops
path.make_simple();
 
// Retread the path
for loc in path {
    // ...
}

Implementations

impl<T: Sized + Clone + Copy + PartialEq + Eq + Hash + Send + Sync + Debug> Path<T>

pub fn from_vec(locs: Vec<T>) -> Self

Create a path from the given locations.

Parameters

locs — The locations to create the path from. The resulting path will have the same locations in the same order as in locs. It must contain at least one location.

Panics

If locs is empty.

Examples

let path = Path::from_vec(vec![3, 2, 1]);
 
let mut iter = path.into_iter();
assert_eq!(Some(3), iter.next());
assert_eq!(Some(2), iter.next());
assert_eq!(Some(1), iter.next());
assert_eq!(None, iter.next());
assert_eq!(None, iter.next());

pub fn starting_at(loc: T) -> Self

Create a path where the first location is loc.

Examples

let path = Path::starting_at(8);

let mut iter = path.into_iter();
assert_eq!(Some(8), iter.next());
assert_eq!(None, iter.next());
assert_eq!(None, iter.next());

pub fn len(&self) -> usize

Return the number of movements in this path.

Examples

let path = Path::from_vec(vec![3, 2, 1]);

assert_eq!(2, path.len());

Cycles don’t shorten the length:

let path = Path::from_vec(vec![3, 2, 1, 2, 3]);

assert_eq!(4, path.len());

pub fn contains(&self, loc: T) -> bool

Test whether this path contains a given location.

Examples

let path = Path::from_vec(vec![3, 2, 1]);

assert!(path.contains(1));
assert!(path.contains(2));
assert!(path.contains(3));
assert!(path.contains(4) == false);

pub fn contains_movement(&self, from: T, to: T) -> bool

Contains a movement from one location to another.

Basically checks if from exists in the path and is immediately followed by to.

Examples

let path = Path::from_vec(vec![3, 2, 1]);

assert!(path.contains_movement(3, 2));
assert!(path.contains_movement(2, 1));
 
// Can't go backwards
assert!(path.contains_movement(1, 2) == false);
assert!(path.contains_movement(2, 3) == false);

// Checks for non-existent points just return false
assert!(path.contains_movement(1, 4) == false);
assert!(path.contains_movement(8, 9) == false);

pub fn push(&mut self, loc: T)

Append a new location to the path.

Examples

let mut path = Path::from_vec(vec![3, 2, 1]);
 
path.push(6);

assert_eq!(3, path[0]);
assert_eq!(2, path[1]);
assert_eq!(1, path[2]);
assert_eq!(6, path[3]);

pub fn is_simple(&self) -> bool

Return whether this path is simple.

A simple path has no loops (no location appears multiple times).

Examples

assert!(Path::from_vec(vec![3, 2, 1]).is_simple());
assert!(Path::from_vec(vec![3, 2, 1, 2]).is_simple() == false);

pub fn make_simple(&mut self)

Make this path into a simple path.

A simple path is a path that has no loops (no repeated locations). This method removes loops by removing all locations between two repeated locations (and then de-duplicating those locations). For example, [4, 1, 5, 1] becomes [4, 1].

If this path is already simple, this method has no effect.

Examples

let mut path = Path::from_vec(vec![3, 2, 1, 2, 3]);
 
path.make_simple();
assert_eq!(Path::from_vec(vec![3]), path);
 
path.make_simple();  // Idempotent
assert_eq!(Path::from_vec(vec![3]), path);

pub fn is_cycle(&self) -> bool

Return where this path is a cycle.

A cycle is a path that starts and ends at the same location. A zero-length path is not considered a cycle.

Examples

assert!(Path::from_vec(vec![1, 2, 1   ]).is_cycle());
assert!(Path::from_vec(vec![1, 2, 2, 1]).is_cycle());
 
assert!(Path::from_vec(vec![1, 2      ]).is_cycle() == false);
assert!(Path::from_vec(vec![1, 2, 2   ]).is_cycle() == false);
assert!(Path::from_vec(vec![1, 2, 1, 2]).is_cycle() == false);
assert!(Path::from_vec(vec![1         ]).is_cycle() == false);

pub fn is_simple_cycle(&self) -> bool

Return whether this path is a simple cycle.

A simple cycle is a path that is a cycle and has no repeated locations, except for the first and last locations.

Examples

assert!(Path::from_vec(vec![1, 2, 1   ]).is_simple_cycle());
 
assert!(Path::from_vec(vec![1, 2, 2, 1]).is_simple_cycle() == false);

See Also

Self::is_simple

Self::is_cycle